security/requirements.txt
opencv-python==4.10.0.84
playsound3==2.0.4
gTTS==2.5.3
deepface==0.0.93
pyautogui==0.9.54
numpy==1.26.4
pillow==10.4.0